Chat_With_Document-demo/Chat_With_Document-demo/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langsmith==0.4.15
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
orjson==3.11.2
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==6.0.0
PyPDF2==3.0.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sentencepiece==0.2.1
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tokenizers==0.21.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.24.0
sentence-transformers
langchain-community
faiss-cpu